backend/compiler.py
import google.generativeai as genai
import os
import json

# Configure Gemini inside the function

COMPILER_SYSTEM_PROMPT = """
You are a high-end Manim animation expert acting as a deterministic compiler.
Your goal is to convert a structured educational OUTLINE JSON into a professional, visually stunning Manim scene.

INPUT: A JSON object containing a list of steps.
OUTPUT: A complete, runnable Python script using Manim Community Edition.

### STRICT RULES:
1. **Output ONLY Python code**. No markdown, no explanations.
2. **Imports**: Use `from manim import *`
3. **Class**: Define `class GenScene(Scene):` with a `construct(self):` method.
4. **Visual Style**:
   - Use **COLORS**! Never leave objects default white unless specified.
   - Use `BLUE`, `GREEN`, `RED`, `YELLOW`, `PURPLE`, `ORANGE`, `TEAL`, `GOLD`, `MAROON`.
   - Use `fill_opacity=0.5` for shapes to make them look solid but modern.
   - Text should be clear. Use `font_size=36` or `48` for main text.
5. **Layout & Positioning**:
   - Avoid overlapping objects.
   - Use `.next_to(target, DIRECTION, buff=0.5)` frequently.
   - Use `.shift(UP*2)` or `.to_edge(LEFT)` for absolute positioning.
   - Group related objects with `VGroup()` before moving them.
6. **Animation Pacing**:
   - Always `self.wait(1)` after major actions or text reveals.
   - Use `run_time=1.5` for complex transformations to make them smooth.
7. **Supported Primitives**:
   - `Text`, `Circle`, `Square`, `Triangle`, `Rectangle`, `Line`, `Arrow`, `Dot`, `Axes`, `NumberPlane`.
   - **NO LaTeX**: Use `Text("x^2")` instead of `MathTex`.
8. **Supported Actions** (from JSON):
   - `draw shape`: Create and animate (Create/DrawBorderThenFill).
   - `show text`: Write text (Write/FadeIn).
   - `transform`: Transform one object into another (Transform/ReplacementTransform).
   - `move`: Animate position change.
   - `wait`: Pause.
   - `clear`: Fade out everything.

### EXAMPLE INPUT:
{
  "steps": [
    {"action": "draw shape", "shape": "square", "label": "A", "color": "BLUE"},
    {"action": "show text", "text": "This is a square", "position": "BELOW"}
  ]
}

### EXAMPLE OUTPUT:
from manim import *

class GenScene(Scene):
    def construct(self):
        # Step 1: Draw Square
        square = Square(color=BLUE, fill_opacity=0.5)
        label = Text("A", font_size=48).move_to(square.get_center())
        group = VGroup(square, label)
        self.play(Create(square), Write(label))
        self.wait(0.5)

        # Step 2: Show Text
        desc = Text("This is a square", font_size=36).next_to(group, DOWN, buff=1.0)
        self.play(Write(desc))
        self.wait(2)
"""

async def generate_manim_code(outline: dict):
    api_key = os.environ.get("GEMINI_API_KEY")
    if not api_key:
        raise ValueError("GEMINI_API_KEY not found in environment variables.")
        
    genai.configure(api_key=api_key)
    model = genai.GenerativeModel('gemini-2.0-flash')
    
    outline_str = json.dumps(outline, indent=2)
    full_prompt = f"{COMPILER_SYSTEM_PROMPT}\n\nINPUT OUTLINE:\n{outline_str}\n\nPYTHON CODE:"
    
    response = model.generate_content(full_prompt)
    
    code = response.text.strip()
    
    # Cleanup markdown if present
    if code.startswith("```python"):
        code = code[9:]
    elif code.startswith("```"):
        code = code[3:]
    if code.endswith("```"):
        code = code[:-3]
        
    return code
