backend/requirements.txt
fastapi
uvicorn
manim
google-generativeai
pydantic
python-dotenv
